+++ test/unit/fuzz-tools.test.js
@@ -399,6 +399,7 @@ import * as cottonMyerV419 from '../../lib/cotton-myer-v419.js';
 import * as friedmanTongueV420 from '../../lib/friedman-tongue-v420.js';
 import * as sunAcCellV421 from '../../lib/sun-ac-cell-v421.js';
 import * as sunAcFlareV422 from '../../lib/sun-ac-flare-v422.js';
+import * as marshOberhuberV423 from '../../lib/marsh-oberhuber-v423.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -743,6 +744,7 @@ const MODULES = {
   'friedman-tongue-v420.js': friedmanTongueV420,
   'sun-ac-cell-v421.js': sunAcCellV421,
   'sun-ac-flare-v422.js': sunAcFlareV422,
+  'marsh-oberhuber-v423.js': marshOberhuberV423,
   'tb-testing.js': tbTesting,
 };